Create UnevenRoundButton component
